requirements.txt
annotated-types==0.7.0
anyio==4.9.0
autopep8==2.3.2
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9
ruff==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
